+++ src/utils/markdown-json/markdownToMarkdownStructure.ts
@@ -11,8 +11,6 @@ import type { MarkdownStructure } from './MarkdownStructure';
  *
  * @param markdown The markdown string to parse.
  * @returns The MarkdownStructure object.
- *
- * @private within the package
  */
 export function markdownToMarkdownStructure(markdown: string): MarkdownStructure {
     const lines = markdown.split('\n');